refactor(ui): replace atomic design with feature-based component structure
Reorganize renderer components from atoms/molecules/organisms/templates/features
into domain-based folders: ui/, canvas/, nodes/, properties/, library/, layout/,
and simulation/. Update all relative and @renderer/ alias imports accordingly.
